// src/lib/api/schemas/GlobalProgressPoint.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Elite API
 * A backend API for https://eliteskyblock.com/ that provides Hypixel Skyblock data.
 * <br><br>
 * Use of this API requires following the [Elite API TOS](https://eliteskyblock.com/apiterms). This API is not affiliated with Hypixel or Mojang.
 * OpenAPI spec version: v1
 */
import type { DictionaryOfStringAndDouble } from './DictionaryOfStringAndDouble';
import type { DictionaryOfStringAndInt64 } from './DictionaryOfStringAndInt64';
import type { GlobalProgressCoverage } from './GlobalProgressCoverage';

export interface GlobalProgressPoint {
	timestamp: bigint;
	updatedAt: bigint;
	isPartial: boolean;
	crops: DictionaryOfStringAndInt64;
	npcPrices: DictionaryOfStringAndDouble;
	cropNpcValues: DictionaryOfStringAndDouble;
	totalCropNpcValue: number;
	npcPricesComplete: boolean;
	pests: DictionaryOfStringAndInt64;
	pestNpcCoinsPerKill: number;
	pestNpcCoins: DictionaryOfStringAndDouble;
	totalPestNpcCoins: number;
	totalNpcValue: number;
	skills: DictionaryOfStringAndDouble;
	cropsCoverage: GlobalProgressCoverage;
	pestsCoverage: GlobalProgressCoverage;
	skillsCoverage: GlobalProgressCoverage;
}